Observer ve factory deseni eklendi

+++ src/main/java/com/example/otelplus/dto/RezervasyonDto.java
@@ -2,6 +2,7 @@ package com.example.otelplus.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -9,7 +10,7 @@ public class RezervasyonDto {
     private Integer rezervasyonId;
     private String otelAdi;
     private String odaTipi;
-    private Double rezervasyonMaliyeti;
+    private BigDecimal rezervasyonMaliyeti;
     private LocalDate girisTarihi;
     private LocalDate cikisTarihi;
 }
